requirements.txt
PyQt6
pyroute2
scapy
python-nmap
netifaces
psutil
